bugfix: Updated BLE device name does not reflect on discovery.
BLE peripheral names are stored in iPhone cache, once iOS connects to the BLE device.
When we discover a previously connected BLE device then its name is retrieved from this cache.
In cases where there is update in name, connection cannot be established with the device.

To fix that, BLE peripheral name is now fetched from advertisement data.

This commit contains another minor change. The device capabilities are now made public so that applications can make use of it.